Refactor: handleDateForm 모듈화
