+++ 01.javascript/todoapp/assets/utils/handleDateForm.js
@@ -1,0 +1,7 @@
+export default function handleDateForm(date) {
+  const dateForm = new Date(date);
+  const year = dateForm.getFullYear();
+  const month = dateForm.getMonth() + 1;
+  const day = dateForm.getDate();
+  return `${year}/${month}/${day}`;
+}
